Add ability to provide custom GraphQL scalars.
GraphQL schemas may have custom scalars, in addition to the base ones
defined in GraphQL spec.

Users of graphql may declare Go types that map to such custom GraphQL
scalars. For example:

	type (
		// DateTime is an ISO-8601 encoded UTC date.
		DateTime struct{ time.Time }

		// GitTimestamp is an ISO-8601 encoded date.
		// Unlike the DateTime type, GitTimestamp is not converted in UTC.
		GitTimestamp struct{ time.Time }

		// URI is an RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI.
		URI struct{ *url.URL }

		// X509Certificate is a valid x509 certificate.
		X509Certificate struct{ *x509.Certificate }
	)

We don't want to expand those struct types when building queries. So,
there needs to be a way to specify which Go types are actually scalars.

This commit creates a way to do that.

This is an interim solution. In the future, the GraphQL client may need
the entire GraphQL schema. Once that's the case, it'll be possible to
know what types are scalars from the schema, so it may be unneded to
specify the Go types that map to GraphQL scalars.