ident: Handle plural form of ID initialism special case.
Unfortunately, for now, handling plural form of ID initialism is a
special case. Plural forms of initialisms are rare and I can't think of
others that would need this handling, so this is good enough for now.
If that changes, we might want to find a more general solution.

Improve internal docs to refer to lower letters as "lower" and upper
letters as "Upper". That helps make it more readable.

Helps shurcooL/githubql#11.