graphql: add methods allowing user-defined query.
Sometimes you want to use a wide variety of queries with the same types.
In this case, the queries generated from the existing types and their tags are
not quite what you want.  So, while writing the queries manually is less fun
than letting the query generator take care of it for you, it can still be
useful!

The operationType enum is removed, because all uses of it coincidentally
disappeared when we moved the query generation calls out from the 'do' func.