query: produce useful message when cycle detected.
Keeping a stack of strings identifing fields we've currently traversed
lets us produce error messages that show the full path that lead us
to a cycle.  (The map is still useful in parallel despite containing
similar data because it's O(1) to check and also because stringifing
types for human readership can be lossy.)

The added test case shows what this error message looks like (including
the representation for anonymous structs).

Signed-off-by: Eric Myhre <hash@exultant.us>