internal/jsonutil: Skip unexported fields.
This change makes jsonutil.UnmarshalGraphQL not consider unexported fields when looking for a matching field.

Document unmarshalValue precondition that v must be addressing and not obtained by the use of unexported fields. That makes it settable, which is a requirement. We want to arrange the internal jsonutil code so that unmarshalValue is never called on an unsettable reflect.Value.

Fixes #36.