remove outdated status note
Packages in this repository are in research and development,
but it's neither active nor early by now. Instead of a fixed
note that may or may not be up-to-date, people can determine
the status more accurately by looking at the code frequency
graph, and API stability by looking at the semantic version.